Initialize redis and azureClient in App
I stedet for å initialisere disse rundt omkring i appen, så flyttes det til å initialiseres i App. Det _kan_ være noe state knyttet til Redisstore som det er greit å beholde.
